fixed some syntactical stuff and moved TabItem to its own type file
