+++ src/components/leaderboard/types.ts
@@ -1,0 +1,7 @@
+export interface TabItem {
+  rank?: number;
+  id: number;
+  name: string;
+  rightSide: React.ReactNode;
+  collapseContent?: React.ReactNode;
+}
