+++ src/containers/volunteerLeaderboard/ducks/selectors.ts
@@ -1,6 +1,6 @@
 import { AsyncRequest, AsyncRequestKinds } from '../../../utils/asyncRequest';
 import { VolunteerLeaderboardItem } from './types';
-import { TabItem } from '../../../components/leaderboard/leaderboardSpace';
+import { TabItem } from '../../../components/leaderboard/types';
 
 export const mapVolunteersToTabItems = (
   items: AsyncRequest<VolunteerLeaderboardItem[], any>,
